requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.1.7
django-appconf==1.0.2
django-choices==1.6.2
django-classy-tags==0.8.0
django-cms==3.6.0
django-compressor==2.2
django-countries==5.3.3
django-filer==1.4.4
django-filter==2.1.0
django-formtools==2.1
django-heroku==0.3.1
django-hvad==1.8.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-polymorphic==2.0.3
django-sass-processor==0.7.2
django-sekizai==0.10.0
django-treebeard==4.3
django-webpack-loader==0.6.0
djangocms-admin-style==1.3.0
djangocms-attributes-field==1.0.0
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-googlemap==1.2.0
djangocms-link==2.3.1
djangocms-picture==2.1.3
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangorestframework==3.9.1
djangorestframework-stubs==0.3.0
easy-thumbnails==2.6
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.2
flake8==3.7.7
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.4.0
libsass==0.17.0
Markdown==3.0.1
mccabe==0.6.1
nbformat==4.4.0
numpy==1.16.2
pandas==0.24.2
Pillow==5.4.1
plotly==3.7.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
rcssmin==1.0.6
requests==2.21.0
retrying==1.3.3
rjsmin==1.0.12
sentry-sdk==0.7.2
six==1.12.0
sorl-thumbnail==12.5.0
text-unidecode==1.2
toml==0.10.0
traitlets==4.3.2
Unidecode==1.0.23
urllib3==1.24.1
webencodings==0.5.1
whitenoise==4.1.2
